// database/prisma-oppulence/generated/zod/inputTypeSchemas/ProjectIncludeSchema.ts
export const ProjectIncludeSchema: z.ZodType<Prisma.ProjectInclude> = z
    .object({
        users: z.union([z.boolean(), z.lazy(() => ProjectUsersFindManyArgsSchema)]).optional(),
        invites: z.union([z.boolean(), z.lazy(() => ProjectInviteFindManyArgsSchema)]).optional(),
        sentEmails: z.union([z.boolean(), z.lazy(() => SentEmailFindManyArgsSchema)]).optional(),
        links: z.union([z.boolean(), z.lazy(() => LinkFindManyArgsSchema)]).optional(),
        domains: z.union([z.boolean(), z.lazy(() => DomainFindManyArgsSchema)]).optional(),
        tags: z.union([z.boolean(), z.lazy(() => TagFindManyArgsSchema)]).optional(),
        programs: z.union([z.boolean(), z.lazy(() => ProgramFindManyArgsSchema)]).optional(),
        invoices: z.union([z.boolean(), z.lazy(() => InvoiceFindManyArgsSchema)]).optional(),
        customers: z.union([z.boolean(), z.lazy(() => CustomerFindManyArgsSchema)]).optional(),
        defaultDomains: z.union([z.boolean(), z.lazy(() => DefaultDomainsFindManyArgsSchema)]).optional(),
        restrictedTokens: z.union([z.boolean(), z.lazy(() => RestrictedTokenFindManyArgsSchema)]).optional(),
        oAuthCodes: z.union([z.boolean(), z.lazy(() => OAuthCodeFindManyArgsSchema)]).optional(),
        integrations: z.union([z.boolean(), z.lazy(() => IntegrationFindManyArgsSchema)]).optional(),
        installedIntegrations: z.union([z.boolean(), z.lazy(() => InstalledIntegrationFindManyArgsSchema)]).optional(),
        webhooks: z.union([z.boolean(), z.lazy(() => WebhookFindManyArgsSchema)]).optional(),
        registeredDomains: z.union([z.boolean(), z.lazy(() => RegisteredDomainFindManyArgsSchema)]).optional(),
        dashboards: z.union([z.boolean(), z.lazy(() => DashboardFindManyArgsSchema)]).optional(),
        utmTemplates: z.union([z.boolean(), z.lazy(() => UtmTemplateFindManyArgsSchema)]).optional(),
        yearInReviews: z.union([z.boolean(), z.lazy(() => YearInReviewFindManyArgsSchema)]).optional(),
        apps: z.union([z.boolean(), z.lazy(() => AppFindManyArgsSchema)]).optional(),
        Inbox: z.union([z.boolean(), z.lazy(() => InboxFindManyArgsSchema)]).optional(),
        Document: z.union([z.boolean(), z.lazy(() => DocumentFindManyArgsSchema)]).optional(),
        CustomerTag: z.union([z.boolean(), z.lazy(() => CustomerTagFindManyArgsSchema)]).optional(),
        ContractTemplate: z.union([z.boolean(), z.lazy(() => ContractTemplateFindManyArgsSchema)]).optional(),
        Report: z.union([z.boolean(), z.lazy(() => ReportFindManyArgsSchema)]).optional(),
        InvoiceTemplate: z.union([z.boolean(), z.lazy(() => InvoiceTemplateFindManyArgsSchema)]).optional(),
        Lead: z.union([z.boolean(), z.lazy(() => LeadFindManyArgsSchema)]).optional(),
        LeadCategory: z.union([z.boolean(), z.lazy(() => LeadCategoryFindManyArgsSchema)]).optional(),
        Pipeline: z.union([z.boolean(), z.lazy(() => PipelineFindManyArgsSchema)]).optional(),
        Deal: z.union([z.boolean(), z.lazy(() => DealFindManyArgsSchema)]).optional(),
        TransactionCategory: z.union([z.boolean(), z.lazy(() => TransactionCategoryFindManyArgsSchema)]).optional(),
        _count: z.union([z.boolean(), z.lazy(() => ProjectCountOutputTypeArgsSchema)]).optional()
    })
    .strict()