// database/prisma-oppulence/generated/zod/outputTypeSchemas/InboxFindManyArgsSchema.ts
export const InboxFindManyArgsSchema: z.ZodType<Prisma.InboxFindManyArgs> = z
    .object({
        select: InboxSelectSchema.optional(),
        include: InboxIncludeSchema.optional(),
        where: InboxWhereInputSchema.optional(),
        orderBy: z.union([InboxOrderByWithRelationInputSchema.array(), InboxOrderByWithRelationInputSchema]).optional(),
        cursor: InboxWhereUniqueInputSchema.optional(),
        take: z.number().optional(),
        skip: z.number().optional(),
        distinct: z.union([InboxScalarFieldEnumSchema, InboxScalarFieldEnumSchema.array()]).optional()
    })
    .strict()